Fix home page 404 error
Make sure `/` returns index.mdx, but `/index` doesn't.
