+++ src/constants/contentlayer.ts
@@ -1,10 +1,16 @@
 import { allDocs, allConfigs } from 'contentlayer/generated';
 export { allDocs } from 'contentlayer/generated';
 
+export const homePageName = 'index';
 export const notFoundPageName = '_not-found';
 export const navLogoName = '_nav-logo';
 export const navLinksName = '_nav-links';
-export const specialFileNames = [notFoundPageName, navLogoName, navLinksName];
+export const specialFileNames = [
+  homePageName,
+  notFoundPageName,
+  navLogoName,
+  navLinksName,
+];
 
 const filteredDocs = allDocs.filter(doc => {
   if (
